# docs/swagger.yaml
basePath: /api/v1
definitions:
  api_v1.CreateAppRequest:
    properties:
      description:
        example: 这是一个测试应用
        type: string
      name:
        example: 测试应用
        type: string
      status:
        example: 1
        type: integer
    required:
    - name
    type: object
...
  github_com_iiwish_lingjian_internal_model.App:
    properties:
      code:
        type: string
      created_at:
        $ref: '#/definitions/github_com_iiwish_lingjian_pkg_utils.CustomTime'
      creator_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        description: 0:禁用 1:启用
        type: integer
      updated_at:
        $ref: '#/definitions/github_com_iiwish_lingjian_pkg_utils.CustomTime'
      updater_id:
        type: integer
    type: object
...
  github_com_iiwish_lingjian_pkg_utils.CustomTime:
    properties:
      time.Time:
        type: string
    type: object
  github_com_iiwish_lingjian_pkg_utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 灵简服务端API文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 灵简服务端API
  version: "1.0"
paths:
  /apps:
    get:
      consumes:
      - application/json
      description: 获取所有应用列表
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 应用ID
        in: header
        name: App-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_iiwish_lingjian_pkg_utils.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_iiwish_lingjian_pkg_utils.Response'
      security:
      - Bearer: []
      summary: 获取应用列表
      tags:
      - Application
    post:
      consumes:
      - application/json
      description: 创建新的应用
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 应用ID
        in: header
        name: App-ID
        required: true
        type: string
      - description: 创建应用请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_v1.CreateAppRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_iiwish_lingjian_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_iiwish_lingjian_internal_model.App'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_iiwish_lingjian_pkg_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_iiwish_lingjian_pkg_utils.Response'
      security:
      - Bearer: []
      summary: 创建应用
      tags:
      - Application
  /apps/{app_id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的应用
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 应用ID
        in: header
        name: App-ID
        required: true
        type: string
      - description: 应用ID
        in: path
        name: app_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_iiwish_lingjian_pkg_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_iiwish_lingjian_pkg_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_iiwish_lingjian_pkg_utils.Response'
      security:
      - Bearer: []
      summary: 删除应用
      tags:
      - Application
    get:
      consumes:
      - application/json
      description: 获取应用详情
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 应用ID
        in: header
        name: App-ID
        required: true
        type: string
      - description: 应用ID
        in: path
        name: app_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_iiwish_lingjian_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_iiwish_lingjian_internal_model.App'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_iiwish_lingjian_pkg_utils.Response'
      security:
      - Bearer: []
      summary: 获取应用详情
      tags:
      - Application
    put:
      consumes:
      - application/json
      description: 更新已存在的应用
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 应用ID
        in: header
        name: App-ID
        required: true
        type: string
      - description: 应用ID
        in: path
        name: app_id
        required: true
        type: integer
      - description: 更新应用请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_v1.CreateAppRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_iiwish_lingjian_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_iiwish_lingjian_internal_model.App'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_iiwish_lingjian_pkg_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_iiwish_lingjian_pkg_utils.Response'
      security:
      - Bearer: []
      summary: 更新应用
      tags:
      - Application
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"